PR fixups - Removed unnecessary ternary - Broke PharmacyOrderMedicationTable into its own component
